pyproject.toml
[project]
name = "homewav2vec2"
version = "0.1.0"
description = "Continued pretraining of wav2vec2-base on long-form Hindi parent–infant home recordings"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1",
    "torchaudio>=2.1",
    "transformers>=4.36",
    "datasets>=2.16",
    "accelerate>=0.25",
    "soundfile>=0.12",
    "numpy>=1.24",
    "pandas>=2.0",
    "pyarrow>=14.0",
    "python-dotenv>=1.0",
    "huggingface-hub>=0.20",
    "pyyaml>=6.0",
    "rich>=13.0",
    "scikit-learn>=1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/homewav2vec2"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
